Add transaction serializers; Use `TransactionSerializer` in `JsonRpcProvider` methods
